* if src is directory, download the directory in local dest path

+++ lib/s3_utils/method.rb
@@ -25,10 +25,27 @@ module S3Utils
       g = Generator.new(src)
 
       if File.directory?(dest)
-        download_path = File.join(dest, File.basename(src))
-        File.open(download_path, 'w') do |f|
-          g.s3_object.read do |chunk|
-            f.write(chunk)
+        if g.s3_object.exists?
+          download_path = File.join(dest, File.basename(src))
+          File.open(download_path, 'w') do |f|
+            g.s3_object.read do |chunk|
+              f.write(chunk)
+            end
+          end
+        else
+          file_objects = g.tree.children(&:reaf?).map(&:object)
+
+          file_objects.each do |obj|
+            base_dir = File.basename(File.dirname(obj.key))
+            obj_name = File.basename(obj.key)
+
+            unless File.exist?(File.join(dest, base_dir))
+              Dir.mkdir(File.join(dest, base_dir))
+            end
+
+            File.open(File.join(dest, base_dir, obj_name), 'w') do |f|
+              obj.read { |chunk| f.write(chunk) }
+            end
           end
         end
       else
